refactor(kubernautagent): split investigator.go into focused files
Extracts RunWorkflowDiscoveryFromRCA+helpers into investigator_discovery.go,
runRCA/retryRCASubmit into investigator_rca.go, runWorkflowSelection and its
helpers into investigator_workflow_selection.go, and the LLM turn loop
(runLLMLoop/chatOrStream/emitToSink/emitCancellationAudit) into
investigator_loop.go. Reduces investigator.go from 1648 to 629 lines with no
behavioral changes; goimports reordered import groups across the package.
