+++ internal/kubernautagent/investigator/investigator.go
@@ -19,16 +19,12 @@ package investigator
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"sync/atomic"
-	"time"
 
 	"github.com/go-logr/logr"
-	"golang.org/x/sync/errgroup"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
-	"github.com/jordigilh/kubernaut/internal/kubernautagent/alignment"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/audit"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/enrichment"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/metrics"
@@ -319,204 +315,6 @@ func (inv *Investigator) RunRCAExtractionFromConversation(ctx context.Context, m
 		return nil, fmt.Errorf("RCA extraction parse: %w", parseErr)
 	}
 	return result, nil
-}
-
-// RunWorkflowDiscoveryFromRCA executes Phase 3 (workflow selection) using a
-// structured RCA result. This reuses the exact autonomous Phase 3 pipeline
-// (BuildPhase1Context + runWorkflowSelection) so that interactive and
-// autonomous workflow discovery produce consistent results.
-//
-// F5/F6 (#1374): When the investigator has an enricher wired, this function
-// resolves enrichment internally (pre-RCA + post-RCA re-enrichment), mirroring
-// the autonomous Investigate() pipeline. The caller-provided enrichData is used
-// as fallback only when the enricher is nil.
-func (inv *Investigator) RunWorkflowDiscoveryFromRCA(ctx context.Context, signal katypes.SignalContext, rcaResult *katypes.InvestigationResult, enrichData *prompt.EnrichmentData, correlationID string) (*katypes.InvestigationResult, error) {
-	if rcaResult == nil {
-		return nil, fmt.Errorf("RunWorkflowDiscoveryFromRCA: rcaResult must not be nil")
-	}
-
-	rcaCopy := *rcaResult
-	rcaResult = &rcaCopy
-
-	inv.pipeline.AnomalyDetector.Reset()
-
-	inv.autoResolveRCATargetAPIVersion(rcaResult, correlationID)
-
-	preKind := signal.ResourceKind
-	signal = SyncSignalFromRCA(signal, rcaResult.RemediationTarget)
-	signal.Namespace = inv.normalizeNamespace(signal.ResourceKind, signal.Namespace)
-
-	inv.logger.Info("RunWorkflowDiscoveryFromRCA: RCA target state",
-		"rca_target_kind", rcaResult.RemediationTarget.Kind,
-		"rca_target_api_version", rcaResult.RemediationTarget.APIVersion,
-		"signal_kind_before", preKind,
-		"signal_kind", signal.ResourceKind,
-		"signal_api_version", signal.ResourceAPIVersion,
-		"signal_namespace", signal.Namespace,
-		"correlation_id", correlationID)
-
-	// F5 (#1374): Resolve enrichment when the enricher is wired, mirroring
-	// the autonomous path (Investigate). Pre-RCA enrichment uses the signal
-	// target; post-RCA re-enrichment uses the RCA-identified target when it
-	// differs (cross-resource RCA).
-	rawEnrichData, enricherWired, hardFailResult := inv.resolveRCAWorkflowDiscoveryEnrichment(ctx, signal, rcaResult, correlationID)
-	if hardFailResult != nil {
-		return hardFailResult, nil
-	}
-	if enricherWired {
-		enrichData = toPromptEnrichment(rawEnrichData)
-	}
-
-	// F6 (#1374): Attach DetectedLabelsJSON to the signal so workflow
-	// discovery tools forward labels to DS catalog queries, activating
-	// GitOps-aware scoring (parity with autonomous path).
-	inv.attachRCADetectedLabelsJSON(&signal, rawEnrichData, correlationID)
-
-	client, modelName, runtimeParams := inv.resolveForPhase(katypes.PhaseWorkflowDiscovery)
-
-	p1Ctx := BuildPhase1Context(rcaResult)
-	workflowResult, err := inv.runWorkflowSelection(ctx, signal, rcaResult.RCASummary, enrichData, p1Ctx, LLMInvocationContext{
-		CorrelationID: correlationID,
-		Client:        client,
-		ModelName:     modelName,
-		RuntimeParams: runtimeParams,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	FinalizeWorkflowResult(workflowResult, signal, rcaResult, rawEnrichData)
-	return workflowResult, nil
-}
-
-// autoResolveRCATargetAPIVersion auto-resolves rcaResult.RemediationTarget.APIVersion
-// when the LLM omitted it, mirroring the autonomous apiVersionValidationGate in
-// runRCA. It mutates rcaResult in place. Uses the REST mapper via
-// scopeResolver: if the kind maps to exactly one API group, the apiVersion
-// can be inferred deterministically without an extra LLM turn.
-func (inv *Investigator) autoResolveRCATargetAPIVersion(rcaResult *katypes.InvestigationResult, correlationID string) {
-	if rcaResult.RemediationTarget.APIVersion != "" || rcaResult.RemediationTarget.Kind == "" || inv.scopeResolver == nil {
-		return
-	}
-	ambiguous, gvrs, err := inv.scopeResolver.IsAmbiguousKind(rcaResult.RemediationTarget.Kind)
-	if err != nil {
-		inv.logger.Error(err, "RunWorkflowDiscoveryFromRCA: IsAmbiguousKind failed, skipping apiVersion auto-resolve",
-			"kind", rcaResult.RemediationTarget.Kind, "correlation_id", correlationID)
-		return
-	}
-	if !ambiguous && len(gvrs) == 1 {
-		rcaResult.RemediationTarget.APIVersion = gvrToAPIVersion(gvrs[0])
-		inv.logger.Info("RunWorkflowDiscoveryFromRCA: auto-resolved apiVersion for non-ambiguous kind",
-			"kind", rcaResult.RemediationTarget.Kind,
-			"api_version", rcaResult.RemediationTarget.APIVersion,
-			"correlation_id", correlationID)
-	}
-}
-
-// resolveRCAWorkflowDiscoveryEnrichment resolves (and, when the RCA target
-// differs from the signal target, re-resolves) enrichment for the interactive
-// workflow-discovery path. enricherWired is false when inv.enricher is nil,
-// signaling that the caller must leave its own enrichData parameter
-// untouched. hardFailResult is non-nil when re-enrichment hard-failed; the
-// caller must return it immediately (it is already fully finalized).
-func (inv *Investigator) resolveRCAWorkflowDiscoveryEnrichment(ctx context.Context, signal katypes.SignalContext, rcaResult *katypes.InvestigationResult, correlationID string) (rawEnrichData *enrichment.EnrichmentResult, enricherWired bool, hardFailResult *katypes.InvestigationResult) {
-	if inv.enricher == nil {
-		return nil, false, nil
-	}
-
-	enrichmentCache := make(map[string]*enrichment.EnrichmentResult)
-	signalKind, signalName, signalNS := ResolveEnrichmentTarget(signal, nil)
-	signalNS = inv.normalizeNamespace(signalKind, signalNS)
-	rawEnrichData = inv.resolveEnrichmentCached(ctx, enrichmentCache, signalKind, signalName, signalNS, signal.IncidentID)
-
-	postRCAKind, postRCAName, postRCANS := ResolveEnrichmentTarget(signal, rcaResult)
-	postRCANS = inv.normalizeNamespace(postRCAKind, postRCANS)
-	if postRCAKind == signalKind && postRCAName == signalName && postRCANS == signalNS {
-		if rawEnrichData != nil && rawEnrichData.TargetResourceDeleted {
-			rcaResult.Warnings = append(rcaResult.Warnings,
-				deletedResourceWarning(signalKind, signalName, signalNS))
-		}
-		return rawEnrichData, true, nil
-	}
-
-	return inv.reEnrichForRCATargetShift(ctx, enrichmentCache, signal, rcaResult, rawEnrichData,
-		signalKind, signalName, postRCAKind, postRCAName, postRCANS, correlationID)
-}
-
-// reEnrichForRCATargetShift re-runs enrichment against the RCA-resolved
-// remediation target when it differs from the signal's original target
-// (e.g. the LLM identified a different owning resource as the true root
-// cause). Returns a non-nil hardFailResult when the re-enrichment
-// hard-fails, in which case rcaResult has already been mutated to request
-// human review with the rca_incomplete reason.
-func (inv *Investigator) reEnrichForRCATargetShift(
-	ctx context.Context,
-	enrichmentCache map[string]*enrichment.EnrichmentResult,
-	signal katypes.SignalContext,
-	rcaResult *katypes.InvestigationResult,
-	rawEnrichData *enrichment.EnrichmentResult,
-	signalKind, signalName, postRCAKind, postRCAName, postRCANS, correlationID string,
-) (*enrichment.EnrichmentResult, bool, *katypes.InvestigationResult) {
-	inv.logger.Info("RunWorkflowDiscoveryFromRCA: re-enriching with RCA target",
-		"signal", signalKind+"/"+signalName,
-		"rca_target", postRCAKind+"/"+postRCAName,
-		"correlation_id", correlationID)
-	reEnriched := inv.resolveEnrichmentCached(ctx, enrichmentCache, postRCAKind, postRCAName, postRCANS, signal.IncidentID)
-
-	if reEnriched != nil && reEnriched.HardFail {
-		inv.logger.Error(reEnriched.OwnerChainError,
-			"RunWorkflowDiscoveryFromRCA: enrichment hard-failed, triggering rca_incomplete",
-			"correlation_id", correlationID)
-		rcaResult.HumanReviewNeeded = true
-		rcaResult.HumanReviewReason = "rca_incomplete"
-		backfillSeverity(rcaResult, signal)
-		attachDetectedLabels(rcaResult, rawEnrichData)
-		InjectRemediationTarget(rcaResult, signal, rawEnrichData)
-		InjectTargetResourceParameters(rcaResult)
-		return rawEnrichData, true, rcaResult
-	}
-
-	if reEnriched == nil {
-		return rawEnrichData, true, nil
-	}
-
-	if reEnriched.TargetResourceDeleted {
-		rcaResult.Warnings = append(rcaResult.Warnings,
-			deletedResourceWarning(postRCAKind, postRCAName, postRCANS))
-	}
-
-	if allLabelDetectionsFailed(reEnriched.DetectedLabels) {
-		inv.logger.Info("RunWorkflowDiscoveryFromRCA: re-enrichment labels all failed, preserving signal-target labels",
-			"rca_target", postRCAKind+"/"+postRCAName, "correlation_id", correlationID)
-		return rawEnrichData, true, nil
-	}
-
-	return reEnriched, true, nil
-}
-
-// attachRCADetectedLabelsJSON marshals rawEnrichData's DetectedLabels onto
-// signal.DetectedLabelsJSON so workflow discovery tools forward labels to DS
-// catalog queries, activating GitOps-aware scoring (parity with the
-// autonomous Investigate path). No-op when rawEnrichData or its labels are
-// unset (which is always the case when the enricher is not wired).
-func (inv *Investigator) attachRCADetectedLabelsJSON(signal *katypes.SignalContext, rawEnrichData *enrichment.EnrichmentResult, correlationID string) {
-	if rawEnrichData == nil || rawEnrichData.DetectedLabels == nil {
-		return
-	}
-	dlJSON, err := json.Marshal(rawEnrichData.DetectedLabels)
-	if err != nil {
-		inv.logger.Error(err, "RunWorkflowDiscoveryFromRCA: failed to marshal detected labels",
-			"correlation_id", correlationID)
-		return
-	}
-	signal.DetectedLabelsJSON = string(dlJSON)
-	dl := rawEnrichData.DetectedLabels
-	trueCount := countTrueLabels(dl.GitOpsManaged, dl.PDBProtected, dl.HPAEnabled,
-		dl.Stateful, dl.HelmManaged, dl.NetworkIsolated, dl.ResourceQuotaConstrained)
-	inv.logger.V(1).Info("RunWorkflowDiscoveryFromRCA: detected labels attached for workflow scoring",
-		"correlation_id", correlationID,
-		"true_label_count", trueCount,
-		"gitops_tool", dl.GitOpsTool)
 }
 
 // Investigate runs the two-invocation investigation and returns the result.
@@ -828,821 +626,4 @@ type LLMInvocationContext struct {
 	Client        llm.Client
 	ModelName     string
 	RuntimeParams llm.RuntimeParams
-}
-
-func (inv *Investigator) runRCA(ctx context.Context, signal katypes.SignalContext, enrichData *prompt.EnrichmentData, llmCtx LLMInvocationContext) (result *katypes.InvestigationResult, retErr error) {
-	correlationID := llmCtx.CorrelationID
-	promptSignal := SignalToPrompt(signal)
-	LogLabelOverrideOrRejection(inv.logger, signal, promptSignal, correlationID, "RCA")
-	systemPrompt, err := inv.builder.RenderInvestigation(promptSignal)
-	if err != nil {
-		return nil, fmt.Errorf("rendering investigation prompt: %w", err)
-	}
-
-	messages := []llm.Message{
-		{Role: "system", Content: systemPrompt},
-		{Role: "user", Content: fmt.Sprintf("Investigate: %s %s in %s — %s", signal.Severity, signal.Name, signal.Namespace, signal.Message)},
-	}
-
-	loopRes, err := inv.runLLMLoop(ctx, messages, katypes.PhaseRCA, llmCtx)
-	if err != nil {
-		return nil, err
-	}
-
-	alignment.NotifyRCAComplete(ctx, messages)
-
-	var content string
-	switch r := loopRes.(type) {
-	case *CancelledResult:
-		cancelledResult := &katypes.InvestigationResult{
-			Cancelled:           true,
-			CancelledPhase:      string(katypes.PhaseRCA),
-			CancelledAtTurn:     r.Turn,
-			AccumulatedMessages: messagesToAuditFormat(r.Messages),
-		}
-		if r.Tokens != nil {
-			s := r.Tokens.Summary()
-			cancelledResult.TokenUsage = &katypes.TokenUsageSummary{
-				PromptTokens:     s.PromptTokens,
-				CompletionTokens: s.CompletionTokens,
-				TotalTokens:      s.TotalTokens,
-			}
-		}
-		return cancelledResult, nil
-	case *ExhaustedResult:
-		return &katypes.InvestigationResult{
-			HumanReviewNeeded: true,
-			Reason:            fmt.Sprintf("%s during RCA (maxTurns=%d)", r.Reason, inv.maxTurns),
-		}, nil
-	case *SubmitResult:
-		content = r.Content
-	case *TextResult:
-		content = r.Content
-	default:
-		content = ""
-	}
-
-	result, parseErr := inv.resultParser.Parse(content)
-	if parseErr != nil {
-		if retried := inv.retryRCASubmit(ctx, content, messages, llmCtx); retried != nil {
-			result = retried
-			parseErr = nil
-		}
-	}
-	if parseErr != nil && ctx.Err() != nil {
-		return &katypes.InvestigationResult{
-			Cancelled:      true,
-			CancelledPhase: string(katypes.PhaseRCA),
-		}, nil
-	}
-	if parseErr != nil {
-		inv.logger.Info("RCA parse failed after retry, treating as summary",
-			"error", parseErr.Error(),
-			"correlation_id", correlationID)
-		return &katypes.InvestigationResult{
-			RCASummary: content,
-		}, nil
-	}
-
-	// Same-kind sentinel validation gate (Issue #847): when the LLM's
-	// remediation_target.kind matches the signal's resource_kind, the LLM may
-	// be targeting the symptom reporter instead of the actual root cause.
-	// Inject a single correction message and re-run. Max 1 retry.
-	result = inv.sameKindValidationGate(ctx, result, signal, messages, llmCtx)
-
-	// Defense-in-depth: RCA phase must never abort the pipeline via
-	// needs_human_review from the parser. Only gate-level exhaustion (#1044)
-	// is a valid RCA abort. Clear parser-set values BEFORE running the
-	// apiVersion gate so the gate's decision is authoritative.
-	// Aligned with HAPI v1.2.1 where needs_human_review is parser-driven in Phase 3.
-	if result.HumanReviewNeeded {
-		inv.logger.Info("clearing parser-set HumanReviewNeeded during RCA (Phase 3 only)",
-			"reason", result.HumanReviewReason,
-			"correlation_id", correlationID)
-		result.HumanReviewNeeded = false
-		result.HumanReviewReason = ""
-	}
-
-	// apiVersion validation gate (Issue #1044): when the remediation target kind
-	// exists in multiple API groups and api_version is missing, the gate retries
-	// once. On exhaustion it sets HumanReviewNeeded=true — a valid RCA abort.
-	// Runs after clearing so its decision is authoritative and not cleared.
-	result = inv.apiVersionValidationGate(ctx, result, messages, llmCtx)
-
-	return result, nil
-}
-
-const maxRCAParseRetries = 1
-
-// retryRCASubmit performs a single correction attempt when the RCA parse fails
-// (e.g. double-serialized JSON that wasn't caught by the unwrap heuristic, or
-// garbage fields). Mirrors retryWorkflowSubmit but scoped to the RCA phase.
-func (inv *Investigator) retryRCASubmit(ctx context.Context, lastContent string, history []llm.Message, llmCtx LLMInvocationContext) *katypes.InvestigationResult {
-	tokens, correlationID, client, modelName, runtimeParams :=
-		llmCtx.Tokens, llmCtx.CorrelationID, llmCtx.Client, llmCtx.ModelName, llmCtx.RuntimeParams
-	submitOnlyTools := []llm.ToolDefinition{
-		{
-			Name:        SubmitResultToolName,
-			Description: "Submit root cause analysis result.",
-			Parameters:  parser.RCAResultSchema(),
-		},
-	}
-
-	correctionMsg := `Your response could not be parsed. You MUST call submit_result with a JSON object like:
-{"root_cause_analysis":{"summary":"...","severity":"critical","signal_name":"SignalName","contributing_factors":["factor1"],"remediation_target":{"kind":"Deployment","name":"resource","namespace":"ns","api_version":"apps/v1"}},"confidence":0.9}
-
-CRITICAL: root_cause_analysis must be a JSON object, NOT a string. Do NOT wrap it in quotes.`
-
-	retryMessages := make([]llm.Message, len(history))
-	copy(retryMessages, history)
-	retryMessages = append(retryMessages,
-		llm.Message{Role: "assistant", Content: lastContent},
-	)
-
-	for attempt := 0; attempt < maxRCAParseRetries; attempt++ {
-		if ctx.Err() != nil {
-			return nil
-		}
-		inv.logger.Info("parse-level retry for RCA submit",
-			"attempt", attempt+1,
-			"max", maxRCAParseRetries,
-			"correlation_id", correlationID)
-
-		retryMessages = append(retryMessages,
-			llm.Message{Role: "user", Content: correctionMsg},
-		)
-
-		retryEvent := audit.NewEvent(audit.EventTypeLLMRequest, correlationID)
-		retryEvent.EventAction = audit.ActionLLMRequest
-		retryEvent.EventOutcome = audit.OutcomeSuccess
-		retryEvent.Data["model"] = modelName
-		retryEvent.Data["prompt_length"] = totalPromptLength(retryMessages)
-		retryEvent.Data["prompt_preview"] = lastUserMessage(retryMessages, 500)
-		retryEvent.Data["retry_attempt"] = attempt + 1
-		retryEvent.Data["retry_max"] = maxRCAParseRetries
-		retryEvent.Data["phase"] = string(katypes.PhaseRCA)
-		retryEvent.Data["retry_reason"] = "rca_parse_correction"
-		audit.StoreBestEffort(ctx, inv.auditStore, retryEvent, inv.auditLog())
-
-		resp, err := inv.chatOrStream(ctx, client, llm.ChatRequest{
-			Messages: retryMessages,
-			Tools:    submitOnlyTools,
-			Options:  llm.ChatOptions{JSONMode: true, OutputSchema: parser.RCAResultSchema()},
-		}, attempt+1, string(katypes.PhaseRCA), modelName, runtimeParams)
-		if err != nil {
-			inv.logger.Error(err, "RCA retry LLM call failed",
-				"correlation_id", correlationID)
-			continue
-		}
-		if tokens != nil {
-			tokens.Add(resp.Usage)
-		}
-
-		emitToSink(ctx, session.EventTypeReasoningDelta, attempt+1, string(katypes.PhaseRCA), map[string]interface{}{
-			"content":       resp.Message.Content,
-			"retry_attempt": attempt + 1,
-		})
-
-		for _, tc := range resp.ToolCalls {
-			if tc.Name == SubmitResultToolName {
-				result, parseErr := inv.resultParser.Parse(tc.Arguments)
-				if parseErr != nil {
-					inv.logger.Error(parseErr, "RCA retry parse still failed",
-						"correlation_id", correlationID)
-					retryMessages = append(retryMessages, resp.Message)
-					continue
-				}
-				inv.logger.Info("RCA retry succeeded",
-					"correlation_id", correlationID)
-				return result
-			}
-		}
-
-		if resp.Message.Content != "" {
-			result, parseErr := inv.resultParser.Parse(resp.Message.Content)
-			if parseErr == nil {
-				inv.logger.Info("RCA retry succeeded from message content",
-					"correlation_id", correlationID)
-				return result
-			}
-		}
-
-		retryMessages = append(retryMessages, resp.Message)
-	}
-	return nil
-}
-
-func (inv *Investigator) runWorkflowSelection(ctx context.Context, signal katypes.SignalContext, rcaSummary string, enrichData *prompt.EnrichmentData, p1Ctx *prompt.Phase1Data, llmCtx LLMInvocationContext) (result *katypes.InvestigationResult, retErr error) {
-	correlationID := llmCtx.CorrelationID
-	// Apply signal label overrides (target_resource_kind / target_resource_name)
-	// before attaching to context. This ensures workflow discovery tools
-	// (list_available_actions, list_workflows) filter by the correct component.
-	// Defense-in-depth for #1064/#1065: even if enrichment resolved a container
-	// kind (e.g. Namespace), the label override corrects it for tool context.
-	overriddenSignal := ApplySignalLabelOverrides(signal)
-	ctx = katypes.WithSignalContext(ctx, overriddenSignal)
-	inv.logger.Info("runWorkflowSelection: post-override signal",
-		"component_gvk", overriddenSignal.ComponentGVK(),
-		"resource_kind", overriddenSignal.ResourceKind,
-		"resource_api_version", overriddenSignal.ResourceAPIVersion,
-		"correlation_id", correlationID)
-
-	wfPromptSignal := SignalToPrompt(signal)
-	LogLabelOverrideOrRejection(inv.logger, signal, wfPromptSignal, correlationID, "workflow selection")
-	systemPrompt, err := inv.builder.RenderWorkflowSelection(prompt.WorkflowSelectionInput{
-		Signal:     wfPromptSignal,
-		RCASummary: rcaSummary,
-		EnrichData: enrichData,
-		Phase1:     p1Ctx,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("rendering workflow selection prompt: %w", err)
-	}
-
-	messages := []llm.Message{
-		{Role: "system", Content: systemPrompt},
-		{Role: "user", Content: fmt.Sprintf("RCA findings: %s\n\nSelect the appropriate remediation workflow.", rcaSummary)},
-	}
-
-	loopRes, err := inv.runLLMLoop(ctx, messages, katypes.PhaseWorkflowDiscovery, llmCtx)
-	if err != nil {
-		return nil, err
-	}
-
-	content, early, err := inv.handleWorkflowSelectionLoopResult(ctx, loopRes, rcaSummary, messages, llmCtx)
-	if early != nil || err != nil {
-		return early, err
-	}
-
-	result, parseErr := inv.resultParser.Parse(content)
-	if parseErr != nil {
-		return inv.workflowSelectionRetryOrHumanReview(ctx, content, messages, rcaSummary, llmCtx, parseErr,
-			fmt.Sprintf("workflow selection: LLM did not produce parseable result: %s", parseErr)), nil
-	}
-
-	if inv.pipeline.CatalogFetcher == nil {
-		return result, nil
-	}
-	return inv.selfCorrectWorkflowSelection(ctx, result, content, messages, rcaSummary, correlationID, llmCtx)
-}
-
-// handleWorkflowSelectionLoopResult classifies the runLLMLoop outcome for
-// PhaseWorkflowDiscovery. When early is non-nil, the caller must return it
-// (and err) immediately without further parsing. Otherwise content holds the
-// tool-call (or text) payload to parse into an InvestigationResult.
-func (inv *Investigator) handleWorkflowSelectionLoopResult(ctx context.Context, loopRes LoopResult, rcaSummary string, messages []llm.Message, llmCtx LLMInvocationContext) (content string, early *katypes.InvestigationResult, err error) {
-	correlationID := llmCtx.CorrelationID
-	switch r := loopRes.(type) {
-	case *CancelledResult:
-		cancelledResult := &katypes.InvestigationResult{
-			RCASummary:          rcaSummary,
-			Cancelled:           true,
-			CancelledPhase:      string(katypes.PhaseWorkflowDiscovery),
-			CancelledAtTurn:     r.Turn,
-			AccumulatedMessages: messagesToAuditFormat(r.Messages),
-		}
-		if r.Tokens != nil {
-			s := r.Tokens.Summary()
-			cancelledResult.TokenUsage = &katypes.TokenUsageSummary{
-				PromptTokens:     s.PromptTokens,
-				CompletionTokens: s.CompletionTokens,
-				TotalTokens:      s.TotalTokens,
-			}
-		}
-		return "", cancelledResult, nil
-	case *ExhaustedResult:
-		return "", &katypes.InvestigationResult{
-			RCASummary:        rcaSummary,
-			HumanReviewNeeded: true,
-			Reason:            fmt.Sprintf("%s during workflow selection (maxTurns=%d)", r.Reason, inv.maxTurns),
-		}, nil
-	case *SubmitNoWorkflowResult:
-		inv.logger.Info("submit_result_no_workflow sentinel: classifying as no_matching_workflows",
-			"correlation_id", correlationID)
-		return "", &katypes.InvestigationResult{
-			RCASummary:        rcaSummary,
-			HumanReviewNeeded: true,
-			HumanReviewReason: "no_matching_workflows",
-			Reason:            "LLM explicitly declined workflow selection via submit_result_no_workflow",
-		}, nil
-	case *SubmitWithWorkflowResult:
-		return r.Content, nil, nil
-	case *SubmitResult:
-		return r.Content, nil, nil
-	case *TextResult:
-		// #760 v2: LLM returned text instead of a tool call. Try parsing
-		// first — the text may contain a valid investigation result (e.g.
-		// problem_resolved or predictive_no_action where no workflow is
-		// expected). Only fall through to parse-level retry when the
-		// content cannot be parsed at all.
-		if _, textErr := inv.resultParser.Parse(r.Content); textErr == nil {
-			inv.logger.Info("workflow selection: parsed text response directly (no tool call)",
-				"correlation_id", correlationID)
-			return r.Content, nil, nil
-		}
-		return "", inv.workflowSelectionRetryOrHumanReview(ctx, r.Content, messages, rcaSummary, llmCtx, nil,
-			"workflow selection: LLM did not use submit tool after retries"), nil
-	}
-	return "", nil, nil
-}
-
-// workflowSelectionRetryOrHumanReview performs one retryWorkflowSubmit
-// attempt and, on exhaustion, classifies the investigation as
-// no_matching_workflows (or cancelled, if the context ended). Shared by the
-// "LLM returned unparseable text" and "LLM's tool payload failed structural
-// parsing" branches of runWorkflowSelection. When parseErr is non-nil the
-// exhaustion is logged at Error level (a real parse failure); otherwise at
-// Info level (the LLM simply never called a submit tool).
-func (inv *Investigator) workflowSelectionRetryOrHumanReview(ctx context.Context, lastContent string, messages []llm.Message, rcaSummary string, llmCtx LLMInvocationContext, parseErr error, failureReason string) *katypes.InvestigationResult {
-	retryResult := inv.retryWorkflowSubmit(ctx, lastContent, messages, rcaSummary, llmCtx)
-	if retryResult != nil {
-		return retryResult
-	}
-	if ctx.Err() != nil {
-		return &katypes.InvestigationResult{
-			RCASummary:     rcaSummary,
-			Cancelled:      true,
-			CancelledPhase: string(katypes.PhaseWorkflowDiscovery),
-		}
-	}
-	if parseErr != nil {
-		inv.logger.Error(parseErr, "workflow selection parse failed after retries, classifying as no_matching_workflows",
-			"correlation_id", llmCtx.CorrelationID)
-	} else {
-		inv.logger.Info("workflow selection: all retries exhausted, classifying as no_matching_workflows",
-			"correlation_id", llmCtx.CorrelationID)
-	}
-	return &katypes.InvestigationResult{
-		RCASummary:        rcaSummary,
-		HumanReviewNeeded: true,
-		HumanReviewReason: "no_matching_workflows",
-		Reason:            failureReason,
-	}
-}
-
-// selfCorrectWorkflowSelection runs the catalog-validated self-correction
-// loop (up to maxSelfCorrectionAttempts) for the parsed workflow-selection
-// result, then applies catalog enrichment and target-alignment checks to the
-// final accepted (or exhausted) result.
-func (inv *Investigator) selfCorrectWorkflowSelection(ctx context.Context, result *katypes.InvestigationResult, content string, messages []llm.Message, rcaSummary, correlationID string, llmCtx LLMInvocationContext) (*katypes.InvestigationResult, error) {
-	validator, fetchErr := inv.pipeline.CatalogFetcher.FetchValidator(ctx)
-	if fetchErr != nil {
-		inv.logger.Error(fetchErr, "workflow catalog unavailable, requiring human review")
-		result.HumanReviewNeeded = true
-		result.HumanReviewReason = "catalog_unavailable"
-		result.Reason = fmt.Sprintf("workflow catalog unavailable: %s", fetchErr)
-		return result, nil
-	}
-
-	attempt := 0
-	correctionFn := func(r *katypes.InvestigationResult, validationErr error) (*katypes.InvestigationResult, error) {
-		attempt++
-		var errStrs []string
-		if validationErr != nil {
-			errStrs = []string{validationErr.Error()}
-		}
-		inv.emitValidationEvent(ctx, attempt, maxSelfCorrectionAttempts, false, errStrs, r.WorkflowID, correlationID)
-
-		correctionMsg, renderErr := inv.renderCorrectionMessage(validationErr, attempt, maxSelfCorrectionAttempts)
-		if renderErr != nil {
-			inv.logger.Error(renderErr, "failed to render validation error template, using fallback")
-			correctionMsg = fmt.Sprintf("Validation failed: %s. Please select a valid workflow.", validationErr)
-		}
-		messages = append(messages, llm.Message{Role: "assistant", Content: content})
-		messages = append(messages, llm.Message{Role: "user", Content: correctionMsg})
-
-		corrLoopRes, corrErr := inv.runLLMLoop(ctx, messages, katypes.PhaseWorkflowDiscovery, llmCtx)
-		if corrErr != nil {
-			return nil, corrErr
-		}
-		switch cr := corrLoopRes.(type) {
-		case *CancelledResult:
-			return nil, context.Canceled
-		case *ExhaustedResult:
-			r.HumanReviewNeeded = true
-			r.Reason = fmt.Sprintf("self-correction: %s", cr.Reason)
-			return r, nil
-		case *SubmitNoWorkflowResult:
-			return &katypes.InvestigationResult{
-				RCASummary:        rcaSummary,
-				HumanReviewNeeded: true,
-				HumanReviewReason: "no_matching_workflows",
-				Reason:            "LLM declined workflow during self-correction via submit_result_no_workflow",
-			}, nil
-		case *SubmitWithWorkflowResult:
-			content = cr.Content
-		case *SubmitResult:
-			content = cr.Content
-		case *TextResult:
-			content = cr.Content
-		}
-		return inv.resultParser.Parse(content)
-	}
-
-	corrected, corrErr := validator.SelfCorrect(result, maxSelfCorrectionAttempts, correctionFn)
-	if corrErr != nil {
-		if errors.Is(corrErr, context.Canceled) || errors.Is(corrErr, context.DeadlineExceeded) {
-			return &katypes.InvestigationResult{
-				RCASummary:     rcaSummary,
-				Cancelled:      true,
-				CancelledPhase: string(katypes.PhaseWorkflowDiscovery),
-			}, nil
-		}
-		return nil, fmt.Errorf("validation self-correction failed: %w", corrErr)
-	}
-	isValid := !corrected.HumanReviewNeeded
-	var finalErrors []string
-	if !isValid {
-		finalErrors = []string{"validation exhausted all attempts"}
-	}
-	inv.emitValidationEvent(ctx, attempt+1, maxSelfCorrectionAttempts, isValid, finalErrors, corrected.WorkflowID, correlationID)
-	enrichFromCatalog(corrected, validator)
-	CheckWorkflowTargetAlignment(ctx, corrected, validator, correlationID, inv.auditStore, inv.logger)
-	return corrected, nil
-}
-
-const maxParseRetries = 2
-
-// retryWorkflowSubmit performs up to maxParseRetries correction attempts when
-// the LLM returns text or unparseable JSON instead of calling a submit tool.
-// Each retry sends a correction message with examples of both submit tools,
-// with only the two submit tools available (prevents re-investigation).
-// Returns non-nil *InvestigationResult on success or nil when retries exhaust.
-func (inv *Investigator) retryWorkflowSubmit(ctx context.Context, lastContent string, history []llm.Message, rcaSummary string, llmCtx LLMInvocationContext) *katypes.InvestigationResult {
-	tokens, correlationID, client, modelName, runtimeParams :=
-		llmCtx.Tokens, llmCtx.CorrelationID, llmCtx.Client, llmCtx.ModelName, llmCtx.RuntimeParams
-	submitOnlyTools := []llm.ToolDefinition{
-		{
-			Name:        SubmitResultWithWorkflowToolName,
-			Description: "Submit investigation result WITH a selected workflow.",
-			Parameters:  parser.WithWorkflowResultSchema(),
-		},
-		{
-			Name:        SubmitResultNoWorkflowToolName,
-			Description: "Submit investigation result when NO matching workflow exists.",
-			Parameters:  parser.NoWorkflowResultSchema(),
-		},
-	}
-
-	correctionTemplate := `Your response could not be parsed. You MUST call one of these tools:
-
-1. If a workflow matches: call submit_result_with_workflow with JSON like:
-   {"root_cause_analysis":{"summary":"..."},"selected_workflow":{"workflow_id":"...","confidence":0.9},"confidence":0.9}
-
-2. If NO workflow matches: call submit_result_no_workflow with JSON like:
-   {"root_cause_analysis":{"summary":"..."},"reasoning":"explanation why no workflow applies"}
-
-Do NOT respond with plain text. You MUST call one of the above tools.`
-
-	retryMessages := make([]llm.Message, len(history))
-	copy(retryMessages, history)
-	retryMessages = append(retryMessages,
-		llm.Message{Role: "assistant", Content: lastContent},
-	)
-
-	for attempt := 0; attempt < maxParseRetries; attempt++ {
-		if ctx.Err() != nil {
-			return nil
-		}
-		inv.logger.Info("parse-level retry for workflow submit",
-			"attempt", attempt+1,
-			"max", maxParseRetries,
-			"correlation_id", correlationID)
-
-		retryMessages = append(retryMessages,
-			llm.Message{Role: "user", Content: correctionTemplate},
-		)
-
-		retryEvent := audit.NewEvent(audit.EventTypeLLMRequest, correlationID)
-		retryEvent.EventAction = audit.ActionLLMRequest
-		retryEvent.EventOutcome = audit.OutcomeSuccess
-		retryEvent.Data["model"] = modelName
-		retryEvent.Data["prompt_length"] = totalPromptLength(retryMessages)
-		retryEvent.Data["prompt_preview"] = lastUserMessage(retryMessages, 500)
-		retryEvent.Data["retry_attempt"] = attempt + 1
-		retryEvent.Data["retry_max"] = maxParseRetries
-		retryEvent.Data["phase"] = string(katypes.PhaseWorkflowDiscovery)
-		retryEvent.Data["retry_reason"] = "parse_level_correction"
-		audit.StoreBestEffort(ctx, inv.auditStore, retryEvent, inv.auditLog())
-
-		resp, err := inv.chatOrStream(ctx, client, llm.ChatRequest{
-			Messages: retryMessages,
-			Tools:    submitOnlyTools,
-			Options:  llm.ChatOptions{JSONMode: true, OutputSchema: parser.InvestigationResultSchema()},
-		}, attempt+1, string(katypes.PhaseWorkflowDiscovery), modelName, runtimeParams)
-		if err != nil {
-			inv.logger.Error(err, "retry LLM call failed",
-				"correlation_id", correlationID)
-			continue
-		}
-		if tokens != nil {
-			tokens.Add(resp.Usage)
-		}
-
-		emitToSink(ctx, session.EventTypeReasoningDelta, attempt+1, string(katypes.PhaseWorkflowDiscovery), map[string]interface{}{
-			"content":       resp.Message.Content,
-			"retry_attempt": attempt + 1,
-		})
-
-		if len(resp.ToolCalls) > 0 {
-			for _, tc := range resp.ToolCalls {
-				switch tc.Name {
-				case SubmitResultNoWorkflowToolName:
-					inv.logger.Info("retry succeeded: submit_result_no_workflow",
-						"correlation_id", correlationID)
-					return &katypes.InvestigationResult{
-						RCASummary:        rcaSummary,
-						HumanReviewNeeded: true,
-						HumanReviewReason: "no_matching_workflows",
-						Reason:            "LLM used submit_result_no_workflow after retry",
-					}
-				case SubmitResultWithWorkflowToolName:
-					inv.logger.Info("retry succeeded: submit_result_with_workflow",
-						"correlation_id", correlationID)
-					result, parseErr := inv.resultParser.Parse(tc.Arguments)
-					if parseErr != nil {
-						inv.logger.Error(parseErr, "retry submit_result_with_workflow parse failed",
-							"correlation_id", correlationID)
-						retryMessages = append(retryMessages, resp.Message)
-						continue
-					}
-					return result
-				}
-			}
-		}
-
-		retryMessages = append(retryMessages, resp.Message)
-	}
-	return nil
-}
-
-func (inv *Investigator) runLLMLoop(ctx context.Context, messages []llm.Message, phase katypes.Phase, llmCtx LLMInvocationContext) (LoopResult, error) {
-	tokens, correlationID, client, modelName, runtimeParams :=
-		llmCtx.Tokens, llmCtx.CorrelationID, llmCtx.Client, llmCtx.ModelName, llmCtx.RuntimeParams
-	toolDefs := inv.toolDefinitionsForPhase(phase)
-	loopStart := time.Now()
-	truncationRetried := false
-	maxTokens := 0
-
-	for turn := 0; turn < inv.maxTurns; turn++ {
-		if ctx.Err() != nil {
-			emitToSink(ctx, session.EventTypeCancelled, turn, string(phase), nil)
-			return &CancelledResult{
-				Messages: messages,
-				Turn:     turn,
-				Phase:    string(phase),
-				Tokens:   tokens,
-			}, nil
-		}
-
-		reqEvent := audit.NewEvent(audit.EventTypeLLMRequest, correlationID)
-		reqEvent.EventAction = audit.ActionLLMRequest
-		reqEvent.EventOutcome = audit.OutcomeSuccess
-		reqEvent.Data["model"] = modelName
-		reqEvent.Data["prompt_length"] = totalPromptLength(messages)
-		reqEvent.Data["prompt_preview"] = lastUserMessage(messages, 500)
-		reqEvent.Data["toolsets_enabled"] = toolNames(toolDefs)
-		reqEvent.Data["messages"] = messagesToAuditFormat(messages)
-		audit.StoreBestEffort(ctx, inv.auditStore, reqEvent, inv.auditLog())
-
-		chatReq := llm.ChatRequest{
-			Messages: messages,
-			Tools:    toolDefs,
-			Options:  llm.ChatOptions{JSONMode: true, OutputSchema: submitResultSchemaForPhase(phase), MaxTokens: maxTokens},
-		}
-		resp, err := inv.chatOrStream(ctx, client, chatReq, turn, string(phase), modelName, runtimeParams)
-		if err != nil {
-			if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
-				emitToSink(ctx, session.EventTypeCancelled, turn, string(phase), nil)
-				return &CancelledResult{
-					Messages: messages,
-					Turn:     turn,
-					Phase:    string(phase),
-					Tokens:   tokens,
-				}, nil
-			}
-			failEvent := audit.NewEvent(audit.EventTypeResponseFailed, correlationID)
-			failEvent.EventAction = audit.ActionResponseFailed
-			failEvent.EventOutcome = audit.OutcomeFailure
-			failEvent.Data["error_message"] = err.Error()
-			failEvent.Data["phase"] = string(phase)
-			failEvent.Data["duration_seconds"] = time.Since(loopStart).Seconds()
-			audit.StoreBestEffort(ctx, inv.auditStore, failEvent, inv.auditLog())
-			return nil, fmt.Errorf("%s LLM call turn %d: %w", phase, turn, err)
-		}
-
-		if tokens != nil {
-			tokens.Add(resp.Usage)
-		}
-
-		respEvent := audit.NewEvent(audit.EventTypeLLMResponse, correlationID)
-		respEvent.EventAction = audit.ActionLLMResponse
-		respEvent.EventOutcome = audit.OutcomeSuccess
-		respEvent.Data["prompt_tokens"] = resp.Usage.PromptTokens
-		respEvent.Data["completion_tokens"] = resp.Usage.CompletionTokens
-		respEvent.Data["total_tokens"] = resp.Usage.TotalTokens
-		respEvent.Data["has_analysis"] = resp.Message.Content != ""
-		respEvent.Data["analysis_length"] = len(resp.Message.Content)
-		respEvent.Data["analysis_preview"] = truncatePreview(resp.Message.Content, 500)
-		respEvent.Data["analysis_full"] = resp.Message.Content
-		respEvent.Data["analysis_content"] = resp.Message.Content
-		respEvent.Data["tool_call_count"] = len(resp.ToolCalls)
-		respEvent.Data["finish_reason"] = resp.FinishReason
-		audit.StoreBestEffort(ctx, inv.auditStore, respEvent, inv.auditLog())
-
-		emitToSink(ctx, session.EventTypeReasoningDelta, turn, string(phase), map[string]interface{}{
-			"content_preview": truncatePreview(resp.Message.Content, 200),
-			"tool_call_count": len(resp.ToolCalls),
-		})
-
-		if len(resp.ToolCalls) > 0 {
-			for _, tc := range resp.ToolCalls {
-				if sr := sentinelResult(tc); sr != nil {
-					inv.logger.Info("sentinel detected",
-						"tool", tc.Name,
-						"phase", string(phase),
-						"correlation_id", correlationID)
-					return sr, nil
-				}
-			}
-
-			assistantMsg := resp.Message
-			assistantMsg.ToolCalls = resp.ToolCalls
-			messages = append(messages, assistantMsg)
-
-			toolResults := make([]string, len(resp.ToolCalls))
-			var g errgroup.Group
-			for i, tc := range resp.ToolCalls {
-				emitToSink(ctx, session.EventTypeToolCallStart, turn, string(phase), map[string]interface{}{
-					"tool_name":  tc.Name,
-					"tool_index": i,
-				})
-				g.Go(func() error {
-					toolResults[i] = inv.executeTool(ctx, tc.Name, json.RawMessage(tc.Arguments))
-					return nil
-				})
-			}
-			_ = g.Wait()
-
-			for i, tc := range resp.ToolCalls {
-				emitToSink(ctx, session.EventTypeToolResult, turn, string(phase), map[string]interface{}{
-					"tool_name":      tc.Name,
-					"tool_index":     i,
-					"result_preview": truncatePreview(toolResults[i], 200),
-				})
-
-				tcEvent := audit.NewEvent(audit.EventTypeLLMToolCall, correlationID)
-				tcEvent.EventAction = audit.ActionToolExecution
-				tcEvent.EventOutcome = audit.OutcomeSuccess
-				tcEvent.Data["tool_call_index"] = i
-				tcEvent.Data["tool_name"] = tc.Name
-				tcEvent.Data["tool_arguments"] = tc.Arguments
-				tcEvent.Data["tool_result"] = toolResults[i]
-				tcEvent.Data["tool_result_preview"] = truncatePreview(toolResults[i], 500)
-				audit.StoreBestEffort(ctx, inv.auditStore, tcEvent, inv.auditLog())
-
-				messages = append(messages, llm.Message{
-					Role:       "tool",
-					Content:    toolResults[i],
-					ToolCallID: tc.ID,
-					ToolName:   tc.Name,
-				})
-			}
-			if inv.pipeline.AnomalyDetector.TotalExceeded() {
-				return &ExhaustedResult{Reason: "tool budget exhausted"}, nil
-			}
-			continue
-		}
-
-		if resp.FinishReason == llm.FinishReasonLength && !truncationRetried {
-			truncationRetried = true
-			maxTokens = escalateMaxTokens(resp.Usage.CompletionTokens)
-			inv.logger.Info("LLM response truncated, retrying with escalated MaxTokens",
-				"phase", string(phase),
-				"escalated_max_tokens", maxTokens,
-				"correlation_id", correlationID)
-
-			truncEvent := audit.NewEvent(audit.EventTypeLLMResponse, correlationID)
-			truncEvent.EventAction = "truncation_detected"
-			truncEvent.EventOutcome = audit.OutcomeFailure
-			truncEvent.Data["has_analysis"] = resp.Message.Content != ""
-			truncEvent.Data["analysis_length"] = len(resp.Message.Content)
-			truncEvent.Data["analysis_preview"] = truncatePreview(resp.Message.Content, 500)
-			truncEvent.Data["finish_reason"] = resp.FinishReason
-			truncEvent.Data["escalated_max_tokens"] = maxTokens
-			truncEvent.Data["truncated_content_length"] = len(resp.Message.Content)
-			audit.StoreBestEffort(ctx, inv.auditStore, truncEvent, inv.auditLog())
-
-			messages = append(messages, resp.Message)
-			messages = append(messages, llm.Message{
-				Role:    "user",
-				Content: "Your previous response was truncated (output token limit reached). Please provide your complete response. Use the submit_result tool to deliver your structured result.",
-			})
-			continue
-		}
-
-		return &TextResult{Content: resp.Message.Content}, nil
-	}
-
-	return &ExhaustedResult{Reason: "max turns exhausted"}, nil
-}
-
-// escalateMaxTokens computes a higher MaxTokens value for truncation recovery.
-// If the truncated response used N completion tokens, we request 2x. Falls back
-// to a default of 8192 if the usage data is unavailable.
-func (inv *Investigator) chatOrStream(ctx context.Context, client llm.Client, req llm.ChatRequest, turn int, phase string, modelName string, runtimeParams llm.RuntimeParams) (llm.ChatResponse, error) {
-	sink := session.EventSinkFromContext(ctx)
-	if sink == nil {
-		inv.logger.Info("chatOrStream: sink is nil, falling back to non-streaming Chat",
-			"turn", turn, "phase", phase,
-			"session_id", session.SessionIDFromContext(ctx))
-		return llm.ChatWithParams(ctx, client, req, runtimeParams)
-	}
-	inv.logger.Info("chatOrStream: sink active, using streaming",
-		"turn", turn, "phase", phase,
-		"session_id", session.SessionIDFromContext(ctx),
-		"sink_ptr", fmt.Sprintf("%p", sink),
-		"diag_sent", diagSendOK.Load(),
-		"diag_dropped", diagSendDrop.Load(),
-		"diag_nil", diagSinkNil.Load())
-
-	temp := runtimeParams.Temperature
-	req.Options.Temperature = &temp
-
-	callCtx := ctx
-	var cancel context.CancelFunc
-	if runtimeParams.TimeoutSeconds > 0 {
-		callCtx, cancel = context.WithTimeout(ctx, time.Duration(runtimeParams.TimeoutSeconds)*time.Second)
-		defer cancel()
-	}
-
-	return client.StreamChat(callCtx, req, func(evt llm.ChatStreamEvent) error {
-		if evt.Delta != "" {
-			emitToSink(ctx, session.EventTypeTokenDelta, turn, phase, map[string]interface{}{
-				"delta": evt.Delta,
-			})
-		}
-		return nil
-	})
-}
-
-// emitToSink sends an InvestigationEvent to the context-carried event sink
-// using non-blocking send semantics. If the sink is nil (no subscriber) or
-// the channel buffer is full, the event is silently dropped. This ensures
-// the investigation loop is never blocked by a slow SSE consumer.
-func emitToSink(ctx context.Context, eventType string, turn int, phase string, data map[string]interface{}) {
-	sink := session.EventSinkFromContext(ctx)
-	if sink == nil {
-		diagSinkNil.Add(1)
-		return
-	}
-	var raw json.RawMessage
-	if data != nil {
-		var err error
-		raw, err = json.Marshal(data)
-		if err != nil {
-			return
-		}
-	}
-	event := session.InvestigationEvent{
-		Type:  eventType,
-		Turn:  turn,
-		Phase: phase,
-		Data:  raw,
-	}
-	select {
-	case sink <- event:
-		diagSendOK.Add(1)
-	default:
-		diagSendDrop.Add(1)
-	}
-}
-
-// emitCancellationAudit emits an investigation-level cancellation event
-// carrying the phase, turn, token usage, and accumulated messages at the point
-// of cancellation. Enriched per COR-2 (token cost attribution), AUD-4
-// (session cross-reference), AUD-6 (messages for forensic RAG), and SEC-1
-// (content cap at 64KB). The context may already be cancelled so we use
-// context.Background() for the audit store call (fire-and-forget per ADR-038).
-func (inv *Investigator) emitCancellationAudit(ctx context.Context, result *katypes.InvestigationResult, correlationID string) {
-	event := audit.NewEvent(audit.EventTypeInvestigationCancelled, correlationID, audit.WithSessionID(session.SessionIDFromContext(ctx)))
-	event.EventAction = audit.ActionInvestigationCancelled
-	event.EventOutcome = audit.OutcomeFailure
-	event.Data["cancelled_phase"] = result.CancelledPhase
-	event.Data["cancelled_at_turn"] = result.CancelledAtTurn
-	if result.TokenUsage != nil {
-		event.Data["total_prompt_tokens"] = result.TokenUsage.PromptTokens
-		event.Data["total_completion_tokens"] = result.TokenUsage.CompletionTokens
-		event.Data["total_tokens"] = result.TokenUsage.TotalTokens
-	}
-	if len(result.AccumulatedMessages) > 0 {
-		if b, err := json.Marshal(result.AccumulatedMessages); err == nil {
-			s := string(b)
-			if len(s) > maxForensicPayloadBytes {
-				s = s[:maxForensicPayloadBytes]
-			}
-			event.Data["accumulated_messages"] = s
-		}
-	}
-	audit.StoreBestEffort(context.Background(), inv.auditStore, event, inv.logger)
 }
