+++ internal/kubernautagent/investigator/investigator_tools.go
@@ -21,10 +21,10 @@ import (
 	"encoding/json"
 
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/alignment"
-	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/parser"
 	"github.com/jordigilh/kubernaut/pkg/kubernautagent/llm"
 	"github.com/jordigilh/kubernaut/pkg/kubernautagent/tools/summarizer"
+	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 )
 
 func escalateMaxTokens(completionTokens int) int {
