+++ internal/kubernautagent/investigator/stream_events_test.go
@@ -19,6 +19,7 @@ package investigator_test
 import (
 	"context"
 	"encoding/json"
+
 	"github.com/go-logr/logr"
 
 	. "github.com/onsi/ginkgo/v2"
@@ -30,8 +31,8 @@ import (
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/parser"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/prompt"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/session"
-	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 	"github.com/jordigilh/kubernaut/pkg/kubernautagent/llm"
+	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 )
 
 func streamTestInvestigator(client llm.Client) *investigator.Investigator {
