+++ internal/kubernautagent/investigator/token_streaming_test.go
@@ -19,9 +19,10 @@ package investigator_test
 import (
 	"context"
 	"encoding/json"
-	"github.com/go-logr/logr"
 	"strings"
 	"sync/atomic"
+
+	"github.com/go-logr/logr"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -32,8 +33,8 @@ import (
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/parser"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/prompt"
 	"github.com/jordigilh/kubernaut/internal/kubernautagent/session"
-	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 	"github.com/jordigilh/kubernaut/pkg/kubernautagent/llm"
+	katypes "github.com/jordigilh/kubernaut/pkg/kubernautagent/types"
 )
 
 // streamTrackingMockClient simulates an LLM that emits known text chunks
